+++ covalent_slurm_plugin/slurm.py
@@ -27,7 +27,7 @@ import sys
 from copy import deepcopy
 from datetime import datetime
 from pathlib import Path
-from typing import Any, Callable, Dict, List, Tuple, Union
+from typing import Any, Callable, Dict, List, Union
 
 import aiofiles
 import asyncssh
@@ -47,18 +47,19 @@ _EXECUTOR_PLUGIN_DEFAULTS = {
     "ssh_key_file": "",
     "cert_file": None,
     "remote_workdir": "covalent-workdir",
+    "create_unique_workdir": False,
     "slurm_path": None,
     "conda_env": None,
-    "cache_dir": str(Path(get_config("dispatcher.cache_dir")).expanduser().resolve()),
     "options": {
         "parsable": "",
     },
-    "poll_freq": 60,
     "use_srun": True,
     "srun_options": {},
     "srun_append": None,
     "prerun_commands": None,
     "postrun_commands": None,
+    "cache_dir": str(Path(get_config("dispatcher.cache_dir")).expanduser().resolve()),
+    "poll_freq": 60,
     "cleanup": True,
 }
 
@@ -75,9 +76,10 @@ class SlurmExecutor(AsyncBaseExecutor):
         cert_file: Certificate file used to authenticate over SSH, if required.
         sshproxy: Dictionary of parameters for sshproxy, namely the "hosts": List[str], "username": str, and "secret": str.
         remote_workdir: Working directory on the remote cluster.
+        create_unique_workdir: Whether to create a unique working (sub)directory for each task.
         slurm_path: Path to the slurm commands if they are not found automatically.
         conda_env: Name of conda environment on which to run the function.
-        cache_dir: Cache directory used by this executor for temporary files.
+        cache_dir: Local cache directory used by this executor for temporary files.
         options: Dictionary of parameters used to build a Slurm submit script.
         prerun_commands: List of shell commands to run before running the pickled function.
         postrun_commands: List of shell commands to run after running the pickled function.
@@ -96,6 +98,7 @@ class SlurmExecutor(AsyncBaseExecutor):
         cert_file: str = None,
         sshproxy: Dict = None,
         remote_workdir: str = None,
+        create_unique_workdir: bool = None,
         slurm_path: str = None,
         conda_env: str = None,
         cache_dir: str = None,
@@ -125,6 +128,12 @@ class SlurmExecutor(AsyncBaseExecutor):
 
         self.remote_workdir = remote_workdir or get_config("executors.slurm.remote_workdir")
 
+        self.create_unique_workdir = (
+            get_config("executors.slurm.create_unique_workdir")
+            if create_unique_workdir is None
+            else create_unique_workdir
+        )
+
         try:
             self.slurm_path = slurm_path or get_config("executors.slurm.slurm_path")
         except KeyError:
@@ -137,6 +146,8 @@ class SlurmExecutor(AsyncBaseExecutor):
 
         cache_dir = cache_dir or get_config("executors.slurm.cache_dir")
         self.cache_dir = str(Path(cache_dir).expanduser().resolve())
+        if not os.path.exists(self.cache_dir):
+            os.makedirs(self.cache_dir)
 
         # To allow passing empty dictionary
         if options is None:
@@ -169,7 +180,7 @@ class SlurmExecutor(AsyncBaseExecutor):
             print("Polling frequency will be increased to the minimum for Slurm: 60 seconds.")
             self.poll_freq = 60
 
-        self.cleanup = cleanup or get_config("executors.slurm.cleanup")
+        self.cleanup = get_config("executors.slurm.cleanup") if cleanup is None else cleanup
 
         # Ensure that the slurm data is parsable
         if "parsable" not in self.options:
@@ -292,6 +303,7 @@ class SlurmExecutor(AsyncBaseExecutor):
         Function to perform cleanup on remote machine
 
         Args:
+            conn: SSH connection object
             remote_func_filename: Function file on remote machine
             remote_slurm_filename: Slurm script file on remote machine
             remote_py_filename: Python script file on remote machine
@@ -311,18 +323,21 @@ class SlurmExecutor(AsyncBaseExecutor):
         await conn.run(f"rm {remote_stderr_filename}")
 
     def _format_submit_script(
-        self,
-        python_version: str,
-        py_filename: str,
+        self, python_version: str, py_filename: str, current_remote_workdir: str
     ) -> str:
         """Create the SLURM that defines the job, uses srun to run the python script.
 
         Args:
             python_version: Python version required by the pickled function.
+            py_filename: Name of the python script.
+            current_remote_workdir: Current working directory on the remote machine.
 
         Returns:
             script: String object containing a script parsable by sbatch.
         """
+
+        # Add chdir to current working directory
+        self.options["chdir"] = current_remote_workdir
 
         # preamble
         slurm_preamble = "#!/bin/bash\n"
@@ -448,6 +463,7 @@ with open("{result_filename}", "wb") as f:
             info_dict: a dictionary containing all necessary parameters needed to query the
                 status of the execution. Required keys in the dictionary are:
                     A string mapping "job_id" to Slurm job ID.
+            conn: SSH connection object.
 
         Returns:
             status: String describing the job status.
@@ -480,6 +496,7 @@ with open("{result_filename}", "wb") as f:
 
         Args:
             job_id: Slurm job ID.
+            conn: SSH connection object.
 
         Returns:
             None
@@ -508,7 +525,7 @@ with open("{result_filename}", "wb") as f:
         Args:
             result_filename: Name of the pickled result file.
             task_results_dir: Directory on the Covalent server where the result will be copied.
-
+            conn: SSH connection object.
         Returns:
             result: Task result.
         """
@@ -547,10 +564,28 @@ with open("{result_filename}", "wb") as f:
         return result, stdout, stderr, exception
 
     async def run(self, function: Callable, args: List, kwargs: Dict, task_metadata: Dict):
+        """Run a function on a remote machine using Slurm.
+
+        Args:
+            function: Function to be executed.
+            args: List of positional arguments to be passed to the function.
+            kwargs: Dictionary of keyword arguments to be passed to the function.
+            task_metadata: Dictionary of metadata associated with the task.
+
+        Returns:
+            result: Result object containing the result of the function execution.
+        """
         dispatch_id = task_metadata["dispatch_id"]
         node_id = task_metadata["node_id"]
         results_dir = task_metadata["results_dir"]
         task_results_dir = os.path.join(results_dir, dispatch_id)
+
+        if self.create_unique_workdir:
+            current_remote_workdir = os.path.join(
+                self.remote_workdir, dispatch_id, "node_" + str(node_id)
+            )
+        else:
+            current_remote_workdir = self.remote_workdir
 
         result_filename = f"result-{dispatch_id}-{node_id}.pkl"
         slurm_filename = f"slurm-{dispatch_id}-{node_id}.sh"
@@ -559,11 +594,11 @@ with open("{result_filename}", "wb") as f:
 
         if "output" not in self.options:
             self.options["output"] = os.path.join(
-                self.remote_workdir, f"stdout-{dispatch_id}-{node_id}.log"
+                current_remote_workdir, f"stdout-{dispatch_id}-{node_id}.log"
             )
         if "error" not in self.options:
             self.options["error"] = os.path.join(
-                self.remote_workdir, f"stderr-{dispatch_id}-{node_id}.log"
+                current_remote_workdir, f"stderr-{dispatch_id}-{node_id}.log"
             )
 
         result = None
@@ -574,11 +609,11 @@ with open("{result_filename}", "wb") as f:
         app_log.debug(f"Python version: {py_version_func}")
 
         # Create the remote directory
-        app_log.debug(f"Creating remote work directory {self.remote_workdir} ...")
-        cmd_mkdir_remote = f"mkdir -p {self.remote_workdir}"
-        proc_mkdir_cache = await conn.run(cmd_mkdir_remote)
+        app_log.debug(f"Creating remote work directory {current_remote_workdir} ...")
+        cmd_mkdir_remote = f"mkdir -p {current_remote_workdir}"
+        proc_mkdir_remote = await conn.run(cmd_mkdir_remote)
 
-        if client_err := proc_mkdir_cache.stderr.strip():
+        if client_err := proc_mkdir_remote.stderr.strip():
             raise RuntimeError(client_err)
 
         async with aiofiles.tempfile.NamedTemporaryFile(dir=self.cache_dir) as temp_f:
@@ -604,12 +639,14 @@ with open("{result_filename}", "wb") as f:
 
         async with aiofiles.tempfile.NamedTemporaryFile(dir=self.cache_dir, mode="w") as temp_h:
             # Format the SLURM submit script, write to file, and copy to remote filesystem
-            slurm_submit_script = self._format_submit_script(py_version_func, py_script_filename)
+            slurm_submit_script = self._format_submit_script(
+                py_version_func, py_script_filename, current_remote_workdir
+            )
             app_log.debug("Writing slurm submit script to tempfile...")
             await temp_h.write(slurm_submit_script)
             await temp_h.flush()
 
-            remote_slurm_filename = os.path.join(self.remote_workdir, slurm_filename)
+            remote_slurm_filename = os.path.join(current_remote_workdir, slurm_filename)
             app_log.debug(f"Copying slurm submit script to remote fs: {remote_slurm_filename} ...")
             await asyncssh.scp(temp_h.name, (conn, remote_slurm_filename))
 
@@ -662,6 +699,14 @@ with open("{result_filename}", "wb") as f:
         return result
 
     async def teardown(self, task_metadata: Dict):
+        """Perform cleanup on remote machine.
+
+        Args:
+            task_metadata: Dictionary of metadata associated with the task.
+
+        Returns:
+            None
+        """
         if self.cleanup:
             try:
                 app_log.debug("Performing cleanup on remote...")